+++ tests/full/vfs/mod.rs
@@ -1,4 +1,0 @@
-mod memory;
-#[cfg(feature = "relaxed-idb")]
-mod relaxed_idb;
-mod sahpool;