+++ tests/main.rs
@@ -1,3 +1,0 @@
-wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_dedicated_worker);
-
-mod full;